Move song metric functionality to new file
